feat: fix image columns
